+++ tests/test_create.py
@@ -6,7 +6,7 @@ from tests.models import Stuff
 class TestCreateTestCase:
     def get_case(self, **kwargs):
         class MockCreateTestCase(CreateAPITestCaseMixin, mocks.MockTestCase):
-            base_name = 'stuff'
+            base_name = kwargs.pop('base_name', 'stuff')
             factory_class = mocks.StuffFactory
             create_data = {"name": "moar stuff"}
 
@@ -26,6 +26,11 @@ class TestCreateTestCase:
         instance = self.get_case(methodName='dummy')
         assert instance.get_create_response()
 
+    def test_get_lookup_from_response(self):
+        instance = self.get_case(methodName='dummy')
+        response = instance.get_create_response()
+        assert instance.get_lookup_from_response(response.data)
+
     def test_test_create(self):
         instance = self.get_case(methodName='dummy')
         instance.setUp()
@@ -34,3 +39,24 @@ class TestCreateTestCase:
         assert created
         assert isinstance(created, Stuff)
         assert response.data['name'] == created.name
+
+        # try again using a different lookup field
+        instance.response_lookup_field = 'name'
+        instance.lookup_field = 'name'
+        response, created = instance.test_create({'name': 'unique stuff'})
+        assert response
+        assert created
+        assert isinstance(created, Stuff)
+        assert response.data['name'] == created.name
+
+    def test_test_create_with_hyperlinkedmodelserializer(self):
+        instance = self.get_case(methodName='dummy', base_name='stuff-linked')
+        instance.setUp()
+        instance.response_lookup_field = 'name'
+        instance.lookup_field = 'name'
+        response, created = instance.test_create({'name': 'moar unique stuff'})
+        assert response
+        assert created
+        assert isinstance(created, Stuff)
+        assert response.data['name'] == created.name
+        assert response.data['url']